chore: addressing review comments
