+++ master/internal/scheduler/default_resource_provider.go
@@ -7,6 +7,8 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/determined-ai/determined/master/pkg/device"
+
 	"github.com/docker/docker/api/types"
 	"github.com/pkg/errors"
 
@@ -19,6 +21,7 @@ import (
 	"github.com/determined-ai/determined/master/pkg/check"
 	cproto "github.com/determined-ai/determined/master/pkg/container"
 	"github.com/determined-ai/determined/master/pkg/model"
+	image "github.com/determined-ai/determined/master/pkg/tasks"
 )
 
 const (
@@ -576,4 +579,38 @@ func toAddresses(proxy string, info types.ContainerJSON) []Address {
 		}
 	}
 	return addresses
+}
+
+// containerAssignment contains information for tasks have been assigned but not yet started.
+type containerAssignment struct {
+	task                  *Task
+	container             *container
+	agent                 *agentState
+	clusterID             string
+	devices               []device.Device
+	harnessPath           string
+	taskContainerDefaults model.TaskContainerDefaultsConfig
+}
+
+// StartTask notifies the agent that the task is ready to start with the provided task spec.
+func (c *containerAssignment) StartTask(spec image.TaskSpec) {
+	handler := c.agent.handler
+	spec.ClusterID = c.clusterID
+	spec.ContainerID = string(c.container.ID())
+	spec.TaskID = string(c.task.ID)
+	spec.HarnessPath = c.harnessPath
+	spec.TaskContainerDefaults = c.taskContainerDefaults
+	spec.Devices = c.devices
+	handler.System().Tell(handler, sproto.StartTaskOnAgent{
+		Task: c.task.handler,
+		StartContainer: aproto.StartContainer{
+			Container: cproto.Container{
+				Parent:  c.task.handler.Address(),
+				ID:      cproto.ID(c.container.id),
+				State:   cproto.Assigned,
+				Devices: c.devices,
+			},
+			Spec: image.ToContainerSpec(spec),
+		},
+	})
 }
