+++ master/static/srv/k8_init_container_entrypoint.sh
